+++ apps/files/src/actions/deleteAction.ts
@@ -12,7 +12,7 @@ import { Permission } from '@nextcloud/files'
 import { loadState } from '@nextcloud/initial-state'
 import { t } from '@nextcloud/l10n'
 import PQueue from 'p-queue'
-import logger from '../logger.ts'
+import { logger } from '../utils/logger.ts'
 import { askConfirmation, canDisconnectOnly, canUnshareOnly, deleteNode, displayName, shouldAskForConfirmation } from './deleteUtils.ts'
 
 // TODO: once the files app is migrated to the new frontend use the import instead:
