+++ apps/files/src/composables/useHotKeys.ts
@@ -7,9 +7,9 @@ import { getFileActions } from '@nextcloud/files'
 import { useHotKey } from '@nextcloud/vue/composables/useHotKey'
 import { dirname } from 'path'
 import { useRoute, useRouter } from 'vue-router/composables'
-import logger from '../logger.ts'
 import { useUserConfigStore } from '../store/userconfig.ts'
 import { executeAction } from '../utils/actionUtils.ts'
+import { logger } from '../utils/logger.ts'
 import { useRouteParameters } from './useRouteParameters.ts'
 
 /**
