+++ apps/files/src/services/ServiceWorker.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 import { generateUrl, getRootUrl } from '@nextcloud/router'
-import logger from '../logger.ts'
+import { logger } from '../utils/logger.ts'
 
 export default () => {
 	if ('serviceWorker' in navigator) {
