+++ apps/files/src/services/Settings.js
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import logger from '../logger.ts'
+import { logger } from '../utils/logger.ts'
 
 export default class Settings {
 	_settings
