+++ apps/files/src/utils/logger.ts
@@ -1,10 +1,11 @@
-/**
+/*!
  * SPDX-FileCopyrightText: 2022 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
+
 import { getLoggerBuilder } from '@nextcloud/logger'
 
-export default getLoggerBuilder()
+export const logger = getLoggerBuilder()
 	.setApp('files')
 	.detectUser()
 	.build()
