+++ apps/files/src/views/ReferenceFileWidget.vue
@@ -68,7 +68,7 @@ import { defineComponent } from 'vue'
 import FileIcon from 'vue-material-design-icons/File.vue'
 import FolderIcon from 'vue-material-design-icons/Folder.vue'
 import { generateFileUrl } from '../../../files_sharing/src/utils/generateUrl.ts'
-import logger from '../logger.ts'
+import { logger } from '../utils/logger.ts'
 
 // see lib/private/Collaboration/Reference/File/FileReferenceProvider.php
 type Ressource = {
